+++ example/lib/tweet_page.dart
@@ -26,11 +26,16 @@ class TweetPage extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
+      backgroundColor: Colors.white,
       appBar: AppBar(
         title: Text(mediaType),
       ),
       body: ListView(
         children: <Widget>[
+          buildHeader("$mediaType TweetEmbed"),
+          buildTweetEmbed(tweetPath),
+          buildHeader("$mediaType Quote TweetEmbed"),
+          buildTweetEmbed(quoteTweetPath),
           buildHeader("$mediaType TweetView"),
           buildTweet(tweetPath),
           buildHeader("$mediaType CompactTweetView"),
@@ -47,11 +52,42 @@ class TweetPage extends StatelessWidget {
   /// Builds a header for a TweetView
   Widget buildHeader(String headerTitle) {
     return Padding(
-      padding: const EdgeInsets.symmetric(vertical: 8.0),
+      padding: const EdgeInsets.symmetric(vertical: 8.0, horizontal: 8.0),
       child: Text(
         headerTitle,
         textAlign: TextAlign.start,
+        style: TextStyle(fontWeight: FontWeight.w700),
       ),
+    );
+  }
+
+  /// Builds a TweetView from a JSON file
+  Widget buildTweetEmbed(String jsonFile) {
+    return FutureBuilder(
+      future: rootBundle.loadString(jsonFile),
+      builder: (BuildContext context, AsyncSnapshot snapshot) {
+        if (snapshot.hasData) {
+          return Container(
+            margin: EdgeInsets.all(15),
+            child: TweetEmbed.fromTweet(
+              Tweet.fromRawJson(
+                snapshot.data,
+              ),
+              darkMode: false,
+              createdDateDisplayFormat: DateFormat("EEE, MMM d, ''yy"),
+            ),
+          );
+        }
+        if (snapshot.hasError) {
+          return Container(
+            child: Text(
+              snapshot.error.toString(),
+            ),
+          );
+        } else {
+          return Container();
+        }
+      },
     );
   }