+++ lib/src/byline.dart
@@ -10,11 +10,13 @@ class Byline extends StatelessWidget {
     this.tweetVM,
     this.viewMode, {
     Key key,
+    this.showDate,
     this.userNameStyle,
     this.userScreenNameStyle,
   }) : super(key: key);
 
   final TweetVM tweetVM;
+  final bool showDate;
   final TextStyle userNameStyle;
   final TextStyle userScreenNameStyle;
   final ViewMode viewMode;
@@ -41,11 +43,17 @@ class Byline extends StatelessWidget {
                 ),
               ],
             ),
-            Text(
-              "@" + tweetVM.userScreenName + " • " + tweetVM.createdAt,
-              textAlign: TextAlign.left,
-              style: userScreenNameStyle,
-            ),
+            (showDate == null || showDate == true)
+                ? Text(
+                    "@" + tweetVM.userScreenName + " • " + tweetVM.createdAt,
+                    textAlign: TextAlign.left,
+                    style: userScreenNameStyle,
+                  )
+                : Text(
+                    "@" + tweetVM.userScreenName,
+                    textAlign: TextAlign.left,
+                    style: userScreenNameStyle,
+                  ),
           ],
         );
         break;
@@ -82,15 +90,17 @@ class Byline extends StatelessWidget {
                       ),
                     ),
                   ),
-                  Padding(
-                    padding: const EdgeInsets.symmetric(horizontal: 4.0),
-                    child: Text(
-                      "• " + tweetVM.createdAt,
-                      style: userScreenNameStyle,
-                      textAlign: TextAlign.start,
-                      softWrap: false,
-                    ),
-                  ),
+                  (showDate == null || showDate == true)
+                      ? Padding(
+                          padding: const EdgeInsets.symmetric(horizontal: 4.0),
+                          child: Text(
+                            "• " + tweetVM.createdAt,
+                            style: userScreenNameStyle,
+                            textAlign: TextAlign.start,
+                            softWrap: false,
+                          ),
+                        )
+                      : Container(),
                 ],
               ),
             ),
@@ -98,6 +108,7 @@ class Byline extends StatelessWidget {
           ],
         );
       default:
+
         /// should never happen
         return Container();
     }